# WIP - disable validation checks during non-debug mode to boost performance